Move system_tests.runTest() and system_tests.verbose_version() to system_tests.BT
