+++ tests/bugfixes/github/test_pr1475_AVIF.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import system_tests
-# test needs system_tests.vv.enable_bmff=1
-vv=system_tests.verbose_version()
+
+# test needs system_tests.BT.vv.enable_bmff=1
 enable_bmff = 'enable_bmff'
+vv=system_tests.BT.verbose_version()
 bSkip = not (enable_bmff in vv and vv[enable_bmff] == '1')
 
 class pr_1475_avif_avif(metaclass=system_tests.CaseMeta):
